# openapi.yaml
openapi: 3.1.0
info:
  title: Collaborative Travel Planner API
  description: |-
    This is the API for the Collaborative Travel Planner application. It allows users to create, manage, and share travel itineraries. The API supports real-time collaboration features through WebSockets.
    Authentication is handled via JWT Bearer tokens obtained from Google OAuth.
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /trips/{tripId}/permissions:
    get:
      summary: Get trip permission matrix
      description: Retrieves the detailed permission matrix for all members of the trip.
      tags:
        - Collaboration
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Permission matrix retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserPermissions'
                  defaultPermissions:
                    $ref: '#/components/schemas/PermissionSet'
        '403':
          description: Access denied - Admin permissions required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Service not found"
        message:
          type: string
          description: Detailed error description
          example: "The requested service could not be found"
      required:
        - error
    # AI Agent Schemas
...
    UserPermissions:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        userName:
          type: string
        role:
          type: string
          enum: [owner, admin, editor, viewer]
        permissions:
          $ref: '#/components/schemas/PermissionSet'
        grantedBy:
          type: string
          format: uuid
        grantedAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
    PermissionSet:
      type: object
      properties:
        canEditTrip:
          type: boolean
          default: false
        canEditItinerary:
          type: boolean
          default: false
        canManageExpenses:
          type: boolean
          default: false
        canCreateVotes:
          type: boolean
          default: false
        canManageMembers:
          type: boolean
          default: false
        canDeleteContent:
          type: boolean
          default: false
        canExportData:
          type: boolean
          default: false
        canManageBookings:
          type: boolean
          default: false
        canAccessAnalytics:
          type: boolean
          default: false
        canModeratChat:
          type: boolean
          default: false
        granularPermissions:
          type: object
          properties:
            destinations:
              type: object
              properties:
                create:
                  type: boolean
                edit:
                  type: boolean
                delete:
                  type: boolean
            activities:
              type: object
              properties:
                create:
                  type: boolean
                edit:
                  type: boolean
                delete:
                  type: boolean
            expenses:
              type: object
              properties:
                viewAll:
                  type: boolean
                editOwn:
                  type: boolean
                editAll:
                  type: boolean
                approveSettlements:
                  type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT